+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 1. [](#improved)
    * Updated `bin/composer.phar` to latest `2.4.4` version [#3627](https://github.com/getgrav/grav/issues/3627)
 1. [](#bugfix)
-   * Get around a bug where messages were not always being set on session during redirect
+   * New `onBeforeSessionStart()` event to be used to store data lost during session regeneration (e.g. login)
    * Don't fail hard if pages recurse with same path
    * Github workflows security hardening [#3624](https://github.com/getgrav/grav/pull/3624)